Add FullnameMixin to separate fullname functionality from RedditBase
