+++ praw/models/reddit/base.py
@@ -1,8 +1,7 @@
 """Provide the RedditBase class."""
 
 from ...compat import string_types, urlparse
-from ...const import API_PATH
-from ...exceptions import ClientException, PRAWException
+from ...exceptions import ClientException
 from ..base import PRAWBase
 
 
@@ -15,18 +14,6 @@ class RedditBase(PRAWBase):
         if not parsed.netloc:
             raise ClientException("Invalid URL: {}".format(url))
         return parsed.path.rstrip("/").split("/")
-
-    @property
-    def fullname(self):
-        """Return the object's fullname.
-
-        A fullname is an object's kind mapping like ``t3`` followed by an
-        underscore and the object's base36 ID, e.g., ``t1_c5s96e0``.
-
-        """
-        return "{}_{}".format(
-            self._reddit._objector.kind(self), self.id
-        )  # pylint: disable=invalid-name
 
     def __eq__(self, other):
         """Return whether the other instance equals the current."""
@@ -77,22 +64,8 @@ class RedditBase(PRAWBase):
         return not self == other
 
     def _fetch(self):
-        if "_info_path" in dir(self):
-            other = self._reddit.get(
-                self._info_path(), params=self._info_params
-            )
-        else:
-            self._info_params["id"] = self.fullname
-            children = self._reddit.get(
-                API_PATH["info"], params=self._info_params
-            ).children
-            if not children:
-                raise PRAWException(
-                    "No {!r} data returned for thing {}".format(
-                        self.__class__.__name__, self.fullname
-                    )
-                )
-            other = children[0]
+        other = self._reddit.get(self._info_path(), params=self._info_params)
+
         self.__dict__.update(other.__dict__)
         self._fetched = True
